+++ src/main/ruby/truffleruby/core/float.rb
@@ -215,7 +215,7 @@ class Float < Numeric
       if ndigits == 0
         round(half: half)
       elsif ndigits < 0
-        truncate.round(ndigits)
+        to_i.round(ndigits, :half => half)
       elsif infinite? or nan?
         self
       else
@@ -229,6 +229,14 @@ class Float < Numeric
           (self * f).round(half: half) / f.to_f
         end
       end
+    end
+  end
+
+  def truncate(ndigits = 0)
+    if positive?
+      floor(ndigits)
+    else
+      ceil(ndigits)
     end
   end