+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Compatibility:
 
 * Implemented `Float#{floor, ceil}` with `ndigits` argument.
 * Implemented `Thread#fetch`.
+* Implemented `Float#truncate` with `ndigits` argument.
 
 * Implemented `-p` CLI option.